refactor: split monolithic errand components into focused modules
Extract header, layout, metadata-band and register-header from
casedata-errand.component.tsx into separate files.
Extract header and layout from support-errand.component.tsx.
Add estate-text utility. Update tab wrappers for new structure.
